+++ slyengine/include/sly/collections/array.hpp
@@ -1,24 +1,24 @@
 template <typename TItemType>
-Array<TItemType>::Array(const TItemType* items, s32 count)
+Array<TItemType>::Array(const TItemType* items, size_t count)
 {
     init(const_cast<TItemType*>(items), count);
 }
 
 template <typename TItemType>
-Array<TItemType>::Array(TItemType* items, s32 count)
+Array<TItemType>::Array(TItemType* items, size_t count)
 {
     init(items, count);
 }
 
 template <typename TItemType>
-void Array<TItemType>::init(TItemType* items, s32 count)
+void Array<TItemType>::init(TItemType* items, size_t count)
 {
     items_ = items;
     count_ = count;
 }
 
 template <typename TItemType>
-void Array<TItemType>::set(s32 index, const TItemType item)
+void Array<TItemType>::set(size_t index, const TItemType item)
 {
     BOUNDS_CHECK(count_, index);
 
@@ -26,7 +26,7 @@ void Array<TItemType>::set(s32 index, const TItemType item)
 }
 
 template <typename TItemType>
-TItemType Array<TItemType>::get(s32 index) const
+TItemType Array<TItemType>::get(size_t index) const
 {
     BOUNDS_CHECK(count_, index);
 
@@ -34,13 +34,13 @@ TItemType Array<TItemType>::get(s32 index) const
 }
 
 template <typename TItemType>       
-s32 Array<TItemType>::count() const
+size_t Array<TItemType>::count() const
 {
     return count_;
 }
 
 template <typename TItemType>
-TItemType& Array<TItemType>::operator[](s32 index)
+TItemType& Array<TItemType>::operator[](size_t index)
 {
     BOUNDS_CHECK(count_, index);
     return items_[index];
@@ -62,5 +62,5 @@ bool_t Array<TItemType>::next(ptr_t* state) const
 template <typename TItemType>
 TItemType Array<TItemType>::read(ptr_t state) const
 {
-    return get((s32)reinterpret_cast<size_t>(state) - 1);
+    return get((size_t)reinterpret_cast<size_t>(state) - 1);
 }
