+++ slyengine/include/sly/collections/linkedlist.h
@@ -15,20 +15,20 @@ namespace sly {
             LinkedList();
             ~LinkedList();
             
-            s32 add(const TItemType item);
-            void addRange(const TItemType* array, s32 count);
+            size_t add(const TItemType item);
+            void addRange(const TItemType* array, size_t count);
 
-            void insert(s32 index, const TItemType item);
-            void remove(s32 index);
+            void insert(size_t index, const TItemType item);
+            void remove(size_t index);
             
-            TItemType get(s32 index) const;
-            void set(s32 index, const TItemType item);
+            TItemType get(size_t index) const;
+            void set(size_t index, const TItemType item);
 
             void clear();
 
-            TItemType& operator[](const s32 index);
+            TItemType& operator[](const size_t index);
 
-            s32 count() const;
+            size_t count() const;
         
         protected:
             virtual TItemType read(ptr_t state) const;
@@ -43,9 +43,9 @@ namespace sly {
             ListNode* head_;
             ListNode* tail_;
 
-            s32 count_;
+            size_t count_;
 
-            ListNode* nodeAt(s32 index) const;
+            ListNode* nodeAt(size_t index) const;
             ListNode* createNode();
             void freeNode(ListNode* node);
     };
