+++ slyengine/include/sly/collections/stack.h
@@ -12,7 +12,7 @@ namespace sly {
             TItemType pop();
             TItemType peek() const;
 
-            s32 count() const;
+            size_t count() const;
 
         protected:
             virtual TItemType read(ptr_t state) const;
