+++ slyengine/include/sly/collections/stack.hpp
@@ -13,7 +13,7 @@ TItemType Stack<TItemType, TCollectionType>::pop()
 }
 
 template <typename TItemType, typename TCollectionType>
-s32 Stack<TItemType, TCollectionType>::count() const
+size_t Stack<TItemType, TCollectionType>::count() const
 {
     return list_.count();
 }
@@ -27,12 +27,12 @@ TItemType Stack<TItemType, TCollectionType>::peek() const
 template <typename TItemType, typename TCollectionType>
 bool_t Stack<TItemType, TCollectionType>::next(ptr_t* state) const
 {
-    *state = reinterpret_cast<ptr_t>(reinterpret_cast<s32>(*state) + 1);
-    return reinterpret_cast<s32>(*state) <= count();
+    *state = reinterpret_cast<ptr_t>(reinterpret_cast<size_t>(*state) + 1);
+    return reinterpret_cast<size_t>(*state) <= count();
 }
 
 template <typename TItemType, typename TCollectionType>
 TItemType Stack<TItemType, TCollectionType>::read(ptr_t state) const
 {
-    return list_.get(count() - reinterpret_cast<s32>(state));
+    return list_.get(count() - reinterpret_cast<size_t>(state));
 }
